Start annotating remaining domains
