+++ ontology/aao.md
@@ -5,6 +5,7 @@ title: Amphibian gross anatomy
 contact:
   email: david.c.blackburn@gmail.com
   label: David Blackburn
+domain: anatomy and development
 homepage: http://github.com/seger/aao
 is_obsolete: true
 replaced_by: uberon
