+++ ontology/fix.md
@@ -11,6 +11,7 @@ contact:
   email: null
   label: chEBI
 description: An ontology of physico-chemical methods and properties.
+domain: chemistry and biochemistry
 homepage: http://www.ebi.ac.uk/chebi
 in_foundry: false
 products:
