+++ ontology/nif_grossanatomy.md
@@ -5,6 +5,7 @@ title: NIF Gross Anatomy
 contact:
   email: smtifahim@gmail.com
   label: Fahim Imam
+domain: anatomy and development
 homepage: http://neuinfo.org/
 is_obsolete: true
 replaced_by: uberon
